debug routing issue

+++ backend/config/cloudinary.js
@@ -13,8 +13,10 @@ const uploadOnCloudinary = async (filePath) => {
     fs.unlinkSync(filePath);
     return uploadResult.secure_url;
   } catch (error) {
-    fs.unlinkSync(filePath);
-    console.log(error);
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
+    throw error;
   }
 };
